pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "prs10"
description = "Commanding and monitoring library for the Rubidium Frequency Standard PRS10 by Stanford Research Systems"
authors = [
  {name = "Taishi Hashimoto", email = "hashimoto.taishi@nipr.ac.jp"},
]
version = "2024.11.01"
requires-python = ">= 3.10"
dependencies = [
    "pyserial"
]

[tool.setuptools]
package-dir = {"" = "src"}

[project.scripts]
prs10stat = "prs10.__main__:main"
